requirements.txt
tavily-python
requests
beautifulsoup4
playwright
fpdf2
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
lxml
